Add Jupyter notebook guide in README and new homework.ipynb file
This update includes the addition of a new Jupyter notebook, Hw7.ipynb, implementing SVD, SVD++, and NMF models with a comparison of their performance based on RMSE scores. The README file has also been updated with a guide on how to summarize a Jupyter notebook. The guide advises on summarising key points like the objective, methods, results, interpretation, and future directions of the notebook.